+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ UNRELEASED
 ==========
 
 * Dropped support for Sphinx < 3.0
+* Added support for alternative parameter names (``arg``, ``argument``, ``parameter``)
 * Fixed ``TypeError`` when formatting a parametrized ``typing.IO`` annotation
 * Fixed data class displaying a return type in its ``__init__()`` method